Resolve formBuilder state bug
Fix a bug where the id's of the objects representing formComponents in the formBuilder state array would change when formComponents of the same type where dropped in Dustbin. The source of the bug is that each time the formComponent from the FieldsTray was dropped into the Dustbin, that same objects reference is what was being stored in the formBuilder state array, the solution was to create a new object that copies the data stored in the dropped object.